major changes in ProfileMap concerning logging of data to backend

+++ backend/travel/models.py
@@ -14,8 +14,8 @@ class BlogPost(models.Model):
     
 class Country(models.Model):
     name = models.CharField(max_length=100)
-    latitude = models.DecimalField(max_digits=9, decimal_places=8)
-    longitude = models.DecimalField(max_digits=9, decimal_places=8)
+    latitude = models.FloatField()
+    longitude = models.FloatField()
     
     
 class UserProfile(models.Model):
